switch to submodule for telemetry plugin
